+++ .utils/library/random.cpp
@@ -1,4 +1,4 @@
-long long int random(long long int min, long long int max){
+long long int random(int min, int max){
   srand(time(0));
   return (rand()%(max-min+1))+min;
 }
